cleaning up sleep handling

+++ .local/lib/status.pl
@@ -36,20 +36,31 @@ if (@ARGV) {
     exit;
 }
 
-my $max   = 15;
-my $cnt   = $max + 1;
-my $lock  = "/tmp/.hikari.lock";
+my $max  = 15;
+my $cnt  = $max + 1;
+my $susp = "/tmp/.hikari.susp";
+my $lock = "/tmp/.hikari.lock";
 while (1) {
     $cnt++;
     if ( !$ENV{"WAYLAND_DISPLAY"} ) {
         exit 0;
     }
     if ( -e $ENV{"IS_LAPTOP"} ) {
-        if ( -e $lock ) {
-            sleep 1;
-            unlink $lock;
-            system("systemctl suspend");
+        if ( system("pidof hikari-unlocker > /dev/null") == 0 ) {
+            if ( -e $lock ) {
+                my $now = `find $lock -cmin +5`;
+                chomp $now;
+                if ($now) {
+                    system("touch $susp");
+                }
+            }
+            if ( -e $susp ) {
+                sleep 1;
+                system("systemctl suspend");
+            }
         }
+        unlink $susp if -e $susp;
+        unlink $lock if -e $lock;
     }
     if ( $cnt >= $max ) {
         system("$status notify &");
